Started setting up basic AI
Added AIController and setup AIModule to be a pluggable AI component.
Moved Player Team and AI Team into prefabs.
Added GameState to hold teams for a level.
Added OtherTeamsEnumerator as a general way to iterate over the other teams in a level.
Exposed Team.Units that only returns the IEnumerable.