Targetting Fixes and start of Win/Lose state
Added ragdollTargetTransform and TargetTransform to Unit so that the correct location can be targeted in certain states.
Changed FollowWorld to have a followUnit rather than a followTransform so that TargetTransform can be used.
Implemented Unit.CanChangeState() to prevent invalid state changes.
Moved setting target to null out of Unit.EndCombat() as it was incorrectly clearing it when MovingToFight after Fighting.
Added ownTeam to GameState.
Setup timed update logic in GameState for checking Win/Loss and Units dropping below a deathPlane.
Added Team.HasUnits().